chore: expose cabinet form types
