+++ src/ui/forms/index.ts
@@ -2,4 +2,4 @@ export { default as ApplianceCabinetForm } from './ApplianceCabinetForm'
 export { default as CargoCabinetForm } from './CargoCabinetForm'
 export { default as CornerCabinetForm } from './CornerCabinetForm'
 export { default as SinkCabinetForm } from './SinkCabinetForm'
-export type { CabinetFormValues, CabinetFormProps } from './CornerCabinetForm'
+export type { CabinetFormValues, CabinetFormProps } from './types'
