+++ src/ui/forms/types.ts
@@ -1,0 +1,12 @@
+export interface CabinetFormValues {
+  width: number
+  height: number
+  depth: number
+  adv?: any
+  hardware?: any
+}
+
+export interface CabinetFormProps {
+  values: CabinetFormValues
+  onChange: (vals: CabinetFormValues) => void
+}
